+++ src/theme/content/content-search-button.ts
@@ -7,6 +7,17 @@
 export default {
   slots: {
     base: '',
+    baseLine: '',
+    label: '',
     trailing: 'hidden lg:flex items-center gap-0.5 ms-auto'
+  },
+  variants: {
+    collapsed: {
+      true: {
+        baseLine: 'ps-[5px] pe-[5px]',
+        label: 'hidden',
+        trailing: 'lg:hidden'
+      }
+    }
   }
 }